pyproject.toml
[project]
name = "jetbrains"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.14.1",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.1.1",
    "google-genai>=1.73.1",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "lxml>=6.1.0",
    "openai>=2.33.0",
    "tavily-python>=0.7.24",
    "torch>=2.11.0",
    "trafilatura>=2.0.0",
    "transformers>=5.6.2",
]
